price-getter/requirements.txt
python-binance==1.0.10
pydantic==1.10.2